connsvc: move keep alive logic into connman
